chore: housekeeping — DRY kickDashboard, close daemon log fd, doc/clarity nits
