+++ lib/init.js
@@ -1,19 +1,10 @@
 'use strict'
 
-const { spawnSync, spawn } = require('child_process')
+const { spawnSync } = require('child_process')
 const path = require('path')
 const { isWindows, installDir } = require('./platform')
 const { bootstrap } = require('./bootstrap')
-
-function kickDashboard() {
-  try {
-    const script = path.join(installDir, 'scripts', 'token-dashboard.py')
-    const p = spawn('python3', [script, '--ensure-daemon'], {
-      detached: true, stdio: 'ignore'
-    })
-    p.unref()
-  } catch (_) { /* never break install/init/update */ }
-}
+const { kickDashboard } = require('./dashboard')
 
 function run(args) {
   bootstrap()
